(BOLT-1103) Add note about builtin tasks requiring the agent
This adds a note to exercise 6 around ensuring that the agent is installed on remote targets before running built in tasks. Users have run into issues trying to run built-in tasks on nodes that don't have an agent installed, so this will warn users that the built in tasks won't work unless the agent is installed. This is a temporary notice until the package and service tasks don't require the agent.